+++ testutil/integration/evm/network/setup.go
@@ -15,7 +15,6 @@ import (
 	feemarkettypes "github.com/cosmos/evm/x/feemarket/types"
 	evmtypes "github.com/cosmos/evm/x/vm/types"
 	"github.com/cosmos/gogoproto/proto"
-	capabilitytypes "github.com/cosmos/ibc-go/modules/capability/types"
 
 	sdkmath "cosmossdk.io/math"
 
@@ -67,7 +66,6 @@ var genesisSetupFunctions = map[string]genSetupFn{
 		// This is handled accordingly on chain and context initialization
 		return genesisState, nil
 	},
-	capabilitytypes.ModuleName: genStateSetter[*capabilitytypes.GenesisState](capabilitytypes.ModuleName),
 }
 
 // genStateSetter is a generic function to set module-specific genesis state